pyproject.toml
[project]
name = "sqltools-mcp"
version = "0.1.0"
description = "A multi-database MCP server supporting MySQL, PostgreSQL, SQL Server, DM8, and SQLite"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "ting", email = "ting@example.com" }
]
keywords = ["mcp", "database", "mysql", "postgresql", "mssql", "dm8", "sqlite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pymysql>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "pymssql>=2.2.0",
    "jaydebeapi>=1.2.0",
    "JPype1>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
sqltools-mcp = "sqltools_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sqltools_mcp"]
